Todo list added

+++ src/App.jsx
@@ -3,19 +3,12 @@ import './App.css'
 import { useAuth } from './firebase/auth.firebase'
 import { useNavigate } from 'react-router-dom';
 import Loader from './components/Loader';
-import { AiOutlinePlus } from "react-icons/ai";
 import { MdDeleteForever } from "react-icons/md";
 import {
-  collection,
-  addDoc,
-  getDocs,
-  where,
-  query,
-  deleteDoc,
-  updateDoc,
-  doc,
+  collection, addDoc, getDocs, where, query, deleteDoc, updateDoc, doc, getDoc,
 } from "firebase/firestore";
-import { db } from './firebase/config.firebase';
+
+import { auth, db } from './firebase/config.firebase';
 import ToastError from './utility/toastify';
 function App() {
   const { authUser, isLoading } = useAuth();
@@ -28,8 +21,12 @@ function App() {
     }
   }, [isLoading, authUser, navigate])
 
+  const [todoTitle, setTodoTitle] = useState("");
   const [todoInput, setTodoInput] = useState("");
+  const [TodoDate, setTodoDate] = useState("");
+  const [Assignuser, setAssignUser] = useState("");
   const [todos, setTodos] = useState([]);
+  const [userList, setUserList] = useState();
   useEffect(() => {
     if (!isLoading && !authUser) {
       router.push("/login");
@@ -44,21 +41,23 @@ function App() {
   }
   const addToDo = async () => {
     try {
-
-
       if (!todoInput.trim().length) {
         EmptyTod();
         return;
       }
-      const docRef = await addDoc(collection(db, "todos"), {
-        owner: authUser.uid,
+      await addDoc(collection(db, "todos"), {
         content: todoInput,
         completed: false,
+        title: todoTitle,
+        description: todoInput,
+        dueDate: TodoDate,
+        assignedUser: Assignuser, // Store assigned user ID here
+        createdBy: authUser.uid,
       });
-
-
       fetchTodos(authUser.uid);
       setTodoInput("");
+      setTodoTitle("");
+      setTodoDate("")
     } catch (error) {
       console.error(error);
     }
@@ -66,34 +65,43 @@ function App() {
 
   const deleteTodo = async (docId) => {
     try {
-      await deleteDoc(doc(db, "todos", docId));
-      fetchTodos(authUser.uid);
+      const DocumentData = await getDoc(doc(db, 'todos', docId))
+      console.log(authUser.uid)
+      const { createdBy, assignedUser } = DocumentData.data()
+      console.log(DocumentData.data())
+
+      if (createdBy == authUser.uid || assignedUser == authUser.email) {
+        await deleteDoc(doc(db, "todos", docId));
+        fetchTodos(authUser.uid);
+        return;
+      }
+      else {
+        ToastError({ message: "Only admin or assigned person can delete this" })
+      }
     } catch (error) {
       console.error(error);
+      ToastError({ message: "Something went Wrong" })
     }
   };
 
   const fetchTodos = async (uid) => {
     try {
-      const q = query(collection(db, "todos"), where("owner", "==", uid));
+      const q = query(collection(db, "todos"));
       const querySnapshot = await getDocs(q);
+
       let data = [];
       querySnapshot.forEach((doc) => {
-
         data.push({ ...doc.data(), id: doc.id });
       });
       setTodos(data);
     } catch (error) {
+      ToastError({message:"Something went wrong file fetching the data"})
       console.error(error);
     }
   };
 
-
-
-
   const markAsCompletedHandler = async (event, docId) => {
     try {
-
       const docRef = doc(db, "todos", docId);
       await updateDoc(docRef, {
         completed: event.target.checked,
@@ -103,6 +111,22 @@ function App() {
       console.error(error);
     }
   };
+
+  useEffect(() => {
+    fetchAllUsers()
+  }, [userList])
+  const fetchAllUsers = async () => {
+    const q = query(collection(db, "users"));
+    const querySnapshot = await getDocs(q);
+    const temp = []
+    querySnapshot.forEach((doc) => {
+      temp.push({
+        ...doc.data()
+      })
+    });
+    setUserList(temp)
+  };
+
 
   const onKeyUp = (event) => {
     if (event.key === "Enter" && todoInput.length > 0) {
@@ -114,62 +138,101 @@ function App() {
     <Loader />
   ) : (
     <main className="">
-
       <div className="max-w-3xl mx-auto mt-10 p-8">
         <div className="bg-white -m-6 p-3 sticky top-0">
           <div className="flex justify-center flex-col items-center">
-            <span className="text-7xl mb-10">📝</span>
-            <h1 className="text-5xl md:text-7xl font-bold">
-              ToooDooo's
+            <h1 className="text-5xl md:text-5xl font-bold">
+            Task Management System
             </h1>
           </div>
-          <div className="flex items-center gap-2 mt-10">
+          <div className="flex items-center gap-4 mt-10 flex-col w-full">
+            <input
+              placeholder={`Write Title of the Todo`}
+              type="text"
+              className="w-full font-semibold placeholder:text-gray-500 border-[2px] border-black h-[60px] grow shadow-sm rounded-md px-4 focus-visible:outline-yellow-400 text-lg transition-all duration-300"
+              autoFocus
+              value={todoTitle}
+              onChange={(e) => setTodoTitle(e.target.value)}
+
+            />
             <input
               placeholder={`👋 Hello Resoulter, What to do Today?`}
               type="text"
-              className="font-semibold placeholder:text-gray-500 border-[2px] border-black h-[60px] grow shadow-sm rounded-md px-4 focus-visible:outline-yellow-400 text-lg transition-all duration-300"
-              autoFocus
+              className="w-full font-semibold placeholder:text-gray-500 border-[2px] border-black h-[60px] grow shadow-sm rounded-md px-4 focus-visible:outline-yellow-400 text-lg transition-all duration-300"
+
               value={todoInput}
               onChange={(e) => setTodoInput(e.target.value)}
-              onKeyUp={onKeyUp}
+
             />
+            <div className='w-full flex justify-center items-center'>
+              <h1 className='w-1/4 mb-4 text-4xl font-semibold leading-none  text-gray-900 dark:text-white'>
+                Due Date
+              </h1>
+              <input
+                placeholder={`👋 Hello Resoulter, What to do Today?`}
+                type="date"
+                className="w-1/4 font-semibold placeholder:text-gray-500 border-[2px] border-black h-[60px] grow shadow-sm rounded-md px-4 focus-visible:outline-yellow-400 text-lg transition-all duration-300"
+                value={TodoDate}
+                onChange={(e) => setTodoDate(e.target.value)}
+                onKeyUp={onKeyUp}
+              />
+            </div>
+            <select id="users" class="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500" onChange={(e) => setAssignUser(e.target.value)}>
+              <option selected value='Z9z2hGnPqlMsqcRvrskzFD4anut1'>Asign Task to someone</option>
+              {
+                userList?.map((user) => {
+                  if (user.email && authUser.email != user.email)
+                    return (
+                      <option value={user.email}>{user.email}</option>
+                    )
+                })
+              }
+            </select>
+
             <button
-              className="w-[60px] h-[60px] rounded-md bg-black flex justify-center items-center cursor-pointer transition-all duration-300 hover:bg-black/[0.8]"
+              className=" w-full h-[60px] text-white rounded-md bg-black flex justify-center items-center cursor-pointer transition-all duration-300 hover:bg-black/[0.8]"
               onClick={addToDo}
             >
-              <AiOutlinePlus size={30} color="#fff" />
+              Add Task
+
             </button>
           </div>
         </div>
         <div className="my-10">
+
+          <div
+            key="{todo.id}"
+            className="flex items-center justify-between mt-4"
+          >
+
+          </div>
+
+
+
+
           {todos.length > 0 &&
             todos.map((todo, index) => (
               <div
                 key={todo.id}
                 className="flex items-center justify-between mt-4"
               >
-                <div className="flex items-center gap-3">
-                  <input
-                    id={`todo-${todo.id}`}
-                    type="checkbox"
-                    className="w-4 h-4 accent-green-400 rounded-lg"
-                    checked={todo.completed}
-                    onChange={(e) =>
-                      markAsCompletedHandler(e, todo.id)
-                    }
-                  />
-                  <label
-                    htmlFor={`todo-${todo.id}`}
-                    className={`font-medium ${todo.completed ? "line-through" : ""
-                      }`}
-                  >
-                    {todo.content}
-                  </label>
-                </div>
+                <ul class="w-full  bg-white rounded-lg shadow divide-gray-200 ">
+                  <li class="px-6 py-4">
+                    <p>
+                      Assigned To: {(authUser.email == todo.assignedUser) ? "You" : todo.assignedUser
+                      }
+                    </p>
+                    <div class="flex justify-between">
+                      <span class="font-semibold text-lg">{todo.title}</span>
+                      <span class="text-gray-500 text-xs"> Due Date:- {todo?.dueDate} + {todo?.assignedUser} </span>
+                    </div>
+                    <p class="text-gray-700 break-words">{todo.description}</p>
 
+                  </li>
+                </ul>
                 <div className="flex items-center gap-3">
                   <MdDeleteForever
-                    size={24}
+                    size={35}
                     className="text-red-400 hover:text-red-600 cursor-pointer"
                     onClick={() => deleteTodo(todo.id)}
                   />
@@ -177,8 +240,8 @@ function App() {
               </div>
             ))}
         </div>
-      </div>
-    </main>
+      </div >
+    </main >
   );
 }
